entire refactor to have snipet description and title as child components fot independent inline edit
